requirements.txt
# LLM-Guided Hierarchical Clustering - Dependencies
# Python 3.9+ required

# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Visualization
matplotlib>=3.4.0

# Machine Learning
scikit-learn>=0.24.0
tslearn>=0.5.2

# Google Gemini API
google-genai>=0.1.0

# Note: python-dotenv is NOT required (API key is in settings.json)
